refactor: move f90wrap_abort implementation to unified library
