+++ f90wrap/scripts/f2py_f90wrap.py
@@ -80,33 +80,10 @@ def main():
     extern char abort_message[ABORT_BUFFER_SIZE];
     void f90wrap_abort_(char *message, int len);
     void f90wrap_abort_int_handler(int signum);
+    void f90wrap_abort__(char *message, int len_message);
 
     #include <stdlib.h>
     #include <string.h>
-
-
-    void f90wrap_abort_(char *message, int len_message)
-    {
-      strncpy(abort_message, message, ABORT_BUFFER_SIZE);
-      abort_message[ABORT_BUFFER_SIZE-1] = '\\0';
-      longjmp(environment_buffer, 0);
-    }
-
-    // copy of f90wrap_abort_ with a second underscore
-    // void (*f90wrap_abort__)(char *, int) = &f90wrap_abort_;
-    void f90wrap_abort__(char *message, int len_message)
-    {
-      strncpy(abort_message, message, ABORT_BUFFER_SIZE);
-      abort_message[ABORT_BUFFER_SIZE-1] = '\\0';
-      longjmp(environment_buffer, 0);
-    }
-
-
-    void f90wrap_abort_int_handler(int signum)
-    {
-      char message[] = "Interrupt occured";
-      f90wrap_abort_(message, strlen(message));
-    }
 
     /* end of custom abort handler  */
     """
